requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.9.2
pyyaml==6.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
minio==7.1.16
joblib==1.3.2
setuptools==65.5.0
